Use the middleware configuration within the plugin
