+++ sdk_contrib/fastify/lib/plugin.js
@@ -1,8 +1,45 @@
 const AWSXray = require('aws-xray-sdk-core');
-const xray = require('./xray');
 
 const xRayFastifyPlugin = (fastify, opts) => {
-  xray(fastify, opts);
+  const defaultOptions = {
+    automaticMode: true,
+    logger: fastify.log,
+  };
+
+  const localOptions = { ...defaultOptions, ...opts };
+
+  if (localOptions.logger) {
+    AWSXray.setLogger(localOptions.logger);
+  } else {
+    AWSXray.setLogger(fastify.log);
+  }
+
+  const segmentName =
+    localOptions.segmentName || exports._internals.createSegmentName();
+  mwUtils.setDefaultName(segmentName);
+
+  if (localOptions.automaticMode) {
+    AWSXray.enableAutomaticMode();
+  } else {
+    AWSXray.enableManualMode();
+  }
+
+  if (localOptions.plugins) {
+    AWSXray.config(localOptions.plugins);
+  }
+
+  if (localOptions.captureAWS) {
+    AWSXray.captureAWS(require('aws-sdk'));
+  }
+
+  if (localOptions.captureHTTP) {
+    AWSXray.captureHTTPsGlobal(require('http'), true);
+    AWSXray.captureHTTPsGlobal(require('https'), true);
+  }
+
+  if (localOptions.capturePromises) {
+    AWSXray.capturePromise();
+  }
 
   const { middleware: mwUtils } = AWSXray;
 
@@ -15,6 +52,8 @@ const xRayFastifyPlugin = (fastify, opts) => {
 
     if (!request.segment) {
       request.segment = segment;
+    } else {
+      fastify.log.warn('Request already has a segment, skipping');
     }
 
     if (AWSXRay.isAutomaticMode()) {
@@ -28,6 +67,8 @@ const xRayFastifyPlugin = (fastify, opts) => {
           next();
         }
       });
+    } else {
+      fastity.log.info('Manual mode, skipping segment');
     }
 
     done();
